// frontend/src/app/model/user/user-info.ts
/**
 * Additional details for a user.
 */
export interface IUserDetails {

    /**
     * ID of the user the info belongs to.
     */
    userId?: number;

    /**
     * Status of the user (if any).
     */
    status?: string;

    /**
     * ID of the user image (if any).
     */
    imageId?: string;

}